feat: Implemented class Individual and crossover and mutation operators for the Pareto front finding genetic algorithm
Created the base genetic algorithm individual and made sure it works with different kinds of problems. Still unsure how to calculate "fitness" of the individuals, so for now we just evaluate all the function values.
Implemented the crossover operator as a whole arithmetic recombination, basically using weighted means of individual code elements.
Implemented the mutation operator as a simple uniform operator which changes a code element value with a new one from that variables domain.