feat: Implemented full Pareto front finding genetic algorithm
Implemented the missing selection operator to select individuals based on the current tournaments Pareto front.
Implemented the genetic algorithm combining the current selection, crossover and mutation operators. Elitism is currently not supported but will most likely be added in the future.
Added plotting of the resulting Pareto fronts to the genetic algorithm. Although the genetic algorithm can take a bit more time than the brute-force one, the resulting graphs are much more dense in their Pareto front region, which could be important in the goal of finding the most optimal MOO problem solutions.